Hash DefId in rustc_span.
